+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConfigChangeReceiverTest.kt
@@ -51,6 +51,7 @@ internal class ConfigChangeReceiverTest {
         appContext,
         resources,
         "",
+        "",
         buildInfo,
         dataDirectory,
         rootDetector,